requirements.txt
django==2.0.4
django-webpack-loader==0.6.0
djangorestframework==3.7.7
django-stronghold==0.3.0
ipython==5.5.0
ipython-genutils==0.2.0
python-dateutil==2.7.2
psycopg2-binary==2.7.4
uWSGI==2.0.17
openpyxl==2.5.2